+++ src/index.ts
@@ -5,6 +5,7 @@ export { default as Container } from '@/package/components/container.vue';
 export { default as Section } from '@/package/components/section.vue';
 export { default as Grid } from '@/package/components/grid.vue';
 export { default as GridItem } from '@/package/components/grid-item.vue';
+export { default as Stack } from '@/package/components/stack.vue';
 export { default as Menu } from '@/package/components/menu.vue';
 export { default as Button } from '@/package/components/button.vue';
 export { default as ButtonGroup } from '@/package/components/button-group.vue';